bug fixes. handle abstract types
