+++ apps/hash-frontend/src/pages/shared/data-type/data-type-constraints/shared/number-input.tsx
@@ -2,6 +2,7 @@ import { Box } from "@mui/material";
 
 export const NumberInput = ({
   id,
+  disabled,
   min,
   max,
   multipleOf,
@@ -10,10 +11,11 @@ export const NumberInput = ({
   width = 120,
 }: {
   id?: string;
+  disabled?: boolean;
   min?: number;
   max?: number;
   multipleOf?: number;
-  onChange: (value: number) => void;
+  onChange: (value: number | undefined) => void;
   value?: number;
   width?: number;
 }) => {
@@ -21,12 +23,21 @@ export const NumberInput = ({
     <Box
       component="input"
       id={id}
+      disabled={disabled}
       step={multipleOf}
       min={min}
       max={max}
       type="number"
       value={value?.toString()}
-      onChange={(event) => onChange(parseInt(event.target.value, 10))}
+      onChange={(event) => {
+        const parsedValue = parseInt(event.target.value, 10);
+
+        if (Number.isNaN(parsedValue)) {
+          onChange(undefined);
+        } else {
+          onChange(parsedValue);
+        }
+      }}
       sx={({ palette }) => ({
         border: `1px solid ${palette.gray[30]}`,
         borderRadius: 1,
